Refactor Category to new directory and add readme
