+++ src/Catalog/Category/CategoryFixturePool.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TddWizard\Fixtures\Catalog;
+namespace TddWizard\Fixtures\Catalog\Category;
 
 use Magento\Catalog\Api\Data\CategoryInterface;
 use function array_values as values;
